Refactor authentication token handling
