+++ janus_server/src/task.rs
@@ -9,8 +9,8 @@ use janus_core::{
         Duration, HpkeAeadId, HpkeConfig, HpkeConfigId, HpkeKdfId, HpkeKemId, HpkePublicKey,
         Interval, Role, TaskId,
     },
+    task::AuthenticationToken,
 };
-use ring::constant_time;
 use serde::{de::Error as _, Deserialize, Deserializer, Serialize, Serializer};
 use std::{
     array::TryFromSliceError,
@@ -164,36 +164,6 @@ impl<const L: usize> TryFrom<&SecretBytes> for VerifyKey<L> {
         Ok(VerifyKey::new(array))
     }
 }
-
-/// An authentication (bearer) token used by aggregators for aggregator-to-aggregator &
-/// collector-to-aggregator authentication.
-#[derive(Clone)]
-pub struct AuthenticationToken(Vec<u8>);
-
-impl From<Vec<u8>> for AuthenticationToken {
-    fn from(token: Vec<u8>) -> Self {
-        Self(token)
-    }
-}
-
-impl AuthenticationToken {
-    /// Returns a view of the aggregator authentication token as a byte slice.
-    pub fn as_bytes(&self) -> &[u8] {
-        &self.0
-    }
-}
-
-impl PartialEq for AuthenticationToken {
-    fn eq(&self, other: &Self) -> bool {
-        // We attempt constant-time comparisons of the token data. Note that this function still
-        // leaks whether the lengths of the tokens are equal -- this is acceptable because we expect
-        // the content of the tokens to provide enough randomness that needs to be guessed even if
-        // the length is known.
-        constant_time::verify_slices_are_equal(&self.0, &other.0).is_ok()
-    }
-}
-
-impl Eq for AuthenticationToken {}
 
 /// The parameters for a PPM task, corresponding to draft-gpew-priv-ppm §4.2.
 #[derive(Clone, Derivative, PartialEq, Eq)]
